NOTICK: Move `InMemoryEntityManagerConfiguration` under testing area
