+++ libs/db/osgi-integration-tests/src/integrationTest/kotlin/net/corda/db/test/osgi/EntitiesInBundlesTest.kt
@@ -5,7 +5,6 @@ import net.corda.db.admin.impl.ClassloaderChangeLog
 import net.corda.db.testkit.DbUtils.getEntityManagerConfiguration
 import net.corda.orm.EntityManagerConfiguration
 import net.corda.orm.EntityManagerFactoryFactory
-import net.corda.orm.impl.InMemoryEntityManagerConfiguration
 import net.corda.test.util.LoggingUtils.emphasise
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterAll
@@ -80,10 +79,7 @@ class EntitiesInBundlesTest {
         private val catId = UUID.randomUUID()
         private val cat = catCtor.newInstance(catId, "Stray", "Tabby", owner)
 
-        private val dbConfig: EntityManagerConfiguration = getEntityManagerConfiguration() ?: run {
-            logger.info("Using in-memory (HSQL) DB".emphasise())
-            InMemoryEntityManagerConfiguration("pets")
-        }
+        private val dbConfig: EntityManagerConfiguration = getEntityManagerConfiguration("pets")
 
         @Suppress("unused")
         @JvmStatic
